schema -> attrs
