+++ dascore/io/tdms/utils.py
@@ -10,7 +10,7 @@ import struct
 import numpy as np
 
 from dascore.core.coords import get_coord
-from dascore.core.schema import PatchAttrs
+from dascore.core.attrs import PatchAttrs
 from dascore.utils.time import to_datetime64, to_timedelta64
 
 DEFAULT_ATTRS = tuple(PatchAttrs.model_fields)
