+++ dascore/io/terra15/utils.py
@@ -6,7 +6,7 @@ from tables.exceptions import NoSuchNodeError
 from dascore.constants import timeable_types
 from dascore.core import Patch
 from dascore.core.coords import get_coord
-from dascore.core.schema import PatchFileSummary
+from dascore.core.attrs import PatchFileSummary
 from dascore.utils.time import to_datetime64, to_timedelta64
 
 # --- Getting format/version
